fix(sync): address #752 review for the data-app-type fix (CLI-8)
- Move create_synced_data_app to a new _sync_data_app module. In
  data_app_service.py it broke the loc-check file-size gate.
- Carry the config's is_disabled flag on the DS create path, a
  regression from the create_config path, which honored it.
- Close the Data Science client through a context manager, so a
  mid-push raise no longer leaks it.
- Delete the DS record if update_config fails after create_app, so a
  failed create leaves no orphan app in the target.
- Drop the stale parameters.id from the create_app call, so the create
  never sends the source project's app id.
- Add a gotchas.md entry for the behavior, with matching notes in
  keboola-expert.md and commands-reference.md, tagged vNEXT.
