+++ src/keboola_agent_cli/services/data_app_service.py
@@ -141,71 +141,6 @@ SECRET_KEY_PATTERN = re.compile(r"^#[A-Za-z][A-Za-z0-9_-]{0,63}$")
 SECRET_OR_PLAIN_KEY_PATTERN = re.compile(r"^#?[A-Za-z][A-Za-z0-9_-]{0,63}$")
 
 
-def create_synced_data_app(
-    storage_client: KeboolaClient,
-    ds_client: DataScienceClient,
-    *,
-    name: str,
-    description: str,
-    type_: str,
-    configuration: dict[str, Any],
-    branch_id: int | None,
-) -> dict[str, Any]:
-    """Create a ``keboola.data-apps`` config together with its Data Science
-    deployment record, carrying the runtime ``type_`` (CLI-8).
-
-    ``sync push`` / ``sync clone`` create configs through the Storage API
-    only (``create_config``). A ``keboola.data-apps`` config has a second
-    half — the Data Science ``/apps`` deployment record — and the runtime
-    type (``python-js`` / ``streamlit`` / ...) lives ONLY on that record,
-    never in the Storage config body. A plain ``create_config`` therefore
-    leaves the platform to lazily mint the DS record under its default type,
-    so a cloned ``python-js`` app deploys as ``streamlit``.
-
-    This routes creation through ``create_app`` (which creates BOTH the DS
-    record with ``type_`` and its Storage config), then fills the full body
-    via ``update_config``. The ``parameters.id`` back-pointer is repointed at
-    the newly assigned app id — the cloned body still carries the SOURCE
-    project's app id, which is meaningless in the target.
-
-    Returns the ``update_config`` response (the Storage config, whose ``id``
-    is the newly assigned config ULID) so the caller's manifest writeback is
-    identical to the ``create_config`` path.
-    """
-    shell = ds_client.create_app(
-        type_=type_,
-        name=name,
-        description="",  # full description goes onto the Storage config below
-        config=configuration,
-        branch_id=branch_id,
-    )
-    app_id = str(shell.get("id", ""))
-    config_id = str(shell.get("configId", ""))
-    if not app_id or not config_id:
-        raise KeboolaApiError(
-            message="POST /apps response missing id or configId",
-            status_code=500,
-            error_code=ErrorCode.API_ERROR,
-            retryable=False,
-        )
-
-    # Repoint the DS back-pointer at the target app; the cloned body still
-    # carries the source project's app id.
-    params = configuration.setdefault("parameters", {})
-    if isinstance(params, dict):
-        params["id"] = app_id
-
-    return storage_client.update_config(
-        component_id=DATA_APP_COMPONENT_ID,
-        config_id=config_id,
-        name=name,
-        description=description,
-        configuration=configuration,
-        change_description="Created via kbagent sync",
-        branch_id=branch_id,
-    )
-
-
 class DataAppService(BaseService):
     """Lifecycle service for Keboola data apps.
